+++ framework-system/framework-management/src/main/java/uk/gov/justice/services/management/shuttering/events/ShutteringRequestedEvent.java
@@ -4,15 +4,25 @@ import uk.gov.justice.services.jmx.api.command.SystemCommand;
 
 import java.time.ZonedDateTime;
 import java.util.Objects;
+import java.util.UUID;
 
 public class ShutteringRequestedEvent {
 
+    private final UUID commandId;
     private final SystemCommand target;
     private final ZonedDateTime shutteringRequestedAt;
 
-    public ShutteringRequestedEvent(final SystemCommand systemCommand, final ZonedDateTime shutteringRequestedAt) {
-        this.target = systemCommand;
+    public ShutteringRequestedEvent(
+            final UUID commandId,
+            final SystemCommand target,
+            final ZonedDateTime shutteringRequestedAt) {
+        this.commandId = commandId;
+        this.target = target;
         this.shutteringRequestedAt = shutteringRequestedAt;
+    }
+
+    public UUID getCommandId() {
+        return commandId;
     }
 
     public SystemCommand getTarget() {
@@ -26,21 +36,23 @@ public class ShutteringRequestedEvent {
     @Override
     public boolean equals(final Object o) {
         if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (!(o instanceof ShutteringRequestedEvent)) return false;
         final ShutteringRequestedEvent that = (ShutteringRequestedEvent) o;
-        return Objects.equals(target, that.target) &&
+        return Objects.equals(commandId, that.commandId) &&
+                Objects.equals(target, that.target) &&
                 Objects.equals(shutteringRequestedAt, that.shutteringRequestedAt);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(target, shutteringRequestedAt);
+        return Objects.hash(commandId, target, shutteringRequestedAt);
     }
 
     @Override
     public String toString() {
         return "ShutteringRequestedEvent{" +
-                "target=" + target +
+                "commandId=" + commandId +
+                ", target=" + target +
                 ", shutteringRequestedAt=" + shutteringRequestedAt +
                 '}';
     }