+++ framework-system/framework-management/src/main/java/uk/gov/justice/services/management/shuttering/observers/shuttering/ShutterCommandApiObserver.java
@@ -29,6 +29,9 @@ public class ShutterCommandApiObserver {
 
         logger.info("Shuttering of Command API complete");
 
-        shutteringRegistry.markShutteringCompleteFor(getClass(), shutteringProcessStartedEvent.getTarget());
+        shutteringRegistry.markShutteringCompleteFor(
+                shutteringProcessStartedEvent.getCommandId(),
+                getClass(),
+                shutteringProcessStartedEvent.getTarget());
     }
 }